#7: use incremental update for ip checksum + move function in file.
