+++ ulb_ipv6.c
@@ -1,4 +1,5 @@
 #include <linux/ipv6.h>
+#include "ulb_checksum.c"
 
 typedef struct in6_addr ip_addr;
